feat: improve savers deposit halt flow
