+++ src/state/apis/swapper/selectors.ts
@@ -13,3 +13,8 @@ export const selectSwapperApiUsdRatesPending = (state: ReduxState) =>
   Object.values(state.swapperApi.queries).some(
     query => query?.endpointName === 'getUsdRates' && query?.status === QueryStatus.pending,
   )
+
+export const selectSwapperApiTradingActivePending = (state: ReduxState) =>
+  Object.values(state.swapperApi.queries).some(
+    query => query?.endpointName === 'getIsTradingActive' && query?.status === QueryStatus.pending,
+  )
